- Fixed compile warning on FreeBSD on 64-bit architecture.   Closes 1493129.
git-svn-id: svn+ssh://svn.code.sf.net/p/opensips/svn/trunk@888 689a6050-402a-0410-94f2-e92a70836424

+++ io_wait.h
@@ -766,7 +766,8 @@ again:
 				   adding smething to the changes array, close() and
 				   applying the changes */
 				LOG(L_INFO, "INFO: io_wait_loop_kqueue: kevent error on "
-							"fd %d: %s [%ld]\n", h->kq_array[r].ident,
+							"fd %u: %s [%ld]\n",
+							(unsigned int)h->kq_array[r].ident,
 							strerror(h->kq_array[r].data),
 							(long)h->kq_array[r].data);
 			}else /* READ/EOF */
